requirements.txt
albumentations==1.1.0
hydra-core==1.3.2
invoke==2.2.0
loguru==0.7.3
Markdown==3.7
matplotlib==3.10.0
numpy==1.24.4
omegaconf==2.3.0
pandas==2.2.3
Pillow==11.1.0
pydantic==2.10.5
PyYAML==6.0.2
scikit_learn==1.6.1
timm==1.0.13
typer==0.15.1
wandb==0.19.4
opencv-python-headless==4.7.0.72
gdown
locust
bentoml
streamlit
